requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.1.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-RESTX==1.3.0
Werkzeug==3.0.1
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
requests==2.31.0
redis==5.0.1
reportlab==4.0.7
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.2.0
bleach==6.1.0
pydantic==2.5.3
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
selenium==4.15.2

